Набросок третьего урока - куб с текстурой, вращаемый мышкой.

+++ src/GLWindow.h
@@ -26,7 +26,7 @@ struct Input
 {
 	uint8_t keyState[256];
 	uint8_t buttonState[3];
-	int16_t cursorPos[2];
+	int     cursorPos[2];
 };
 
 // создание окна с поддержкой OpenGL
@@ -57,8 +57,8 @@ bool InputIsButtonDown(uint8_t button);
 bool InputIsButtonClick(uint8_t button);
 
 // получить текущее положение курсора
-void InputGetCursorPos(int16_t *x, int16_t *y);
+void InputGetCursorPos(int *x, int *y);
 // задать положение курсора
-void InputSetCursorPos(int16_t x, int16_t y);
+void InputSetCursorPos(int x, int y);
 
 #endif /* GLWINDOW_H */
